Cleanup & test

+++ source/shogle/render/shaders/generic2d.hpp
@@ -31,7 +31,7 @@ public:
 
   inline generic2d& bind_texture(const gl::texture& tex) {
     tex.bind(_sprite_sampler);
-    set_uniform(_texture_unif, _sprite_sampler);
+    set_uniform(_texture_unif, (int)_sprite_sampler);
     return *this;
   }
